# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QualificationEvidenceGenerationStatus = typeof QualificationEvidenceGenerationStatus[keyof typeof QualificationEvidenceGenerationStatus];


export const QualificationEvidenceGenerationStatus = {
  COMPLETE: 'COMPLETE',
  CORRECTED: 'CORRECTED',
  FALLBACK: 'FALLBACK',
} as const;

export type EvidenceItemEvidenceState = typeof EvidenceItemEvidenceState[keyof typeof EvidenceItemEvidenceState];


export const EvidenceItemEvidenceState = {
  CONFIRMED: 'CONFIRMED',
  SUPPORTED: 'SUPPORTED',
  EXPLICIT: 'EXPLICIT',
  UNKNOWN: 'UNKNOWN',
  INFERRED: 'INFERRED',
  CONTRADICTED: 'CONTRADICTED',
  ABSENT: 'ABSENT',
} as const;

export type EvidenceItemAssessmentType = typeof EvidenceItemAssessmentType[keyof typeof EvidenceItemAssessmentType];


export const EvidenceItemAssessmentType = {
  COMPANY_FACT: 'COMPANY_FACT',
  COMPANY: 'COMPANY',
  CAPABILITY: 'CAPABILITY',
  WORKFLOW: 'WORKFLOW',
  SAMPLE: 'SAMPLE',
  SAMPLE_TYPE: 'SAMPLE_TYPE',
  PROGRAM: 'PROGRAM',
  MODALITY: 'MODALITY',
  QUESTION: 'QUESTION',
  OPPORTUNITY: 'OPPORTUNITY',
  FIT: 'FIT',
  LIMITATION: 'LIMITATION',
  SOURCE: 'SOURCE',
  UNKNOWN: 'UNKNOWN',
} as const;

export type EvidenceItemProvenanceType = typeof EvidenceItemProvenanceType[keyof typeof EvidenceItemProvenanceType];


export const EvidenceItemProvenanceType = {
  PUBLIC_SOURCE: 'PUBLIC_SOURCE',
  CONFIRMED_ACCOUNT: 'CONFIRMED_ACCOUNT',
} as const;

export interface EvidenceItem {
  evidenceState: EvidenceItemEvidenceState;
  assessmentType: EvidenceItemAssessmentType;
  /** @minLength 1 */
  claim: string;
  /**
     * @nullable
     * @pattern ^https?://
     */
  sourceUrl: string | null;
  basisFacts: string[];
  /** @items.pattern ^https?:// */
  basisSourceUrls: string[];
  /** @nullable */
  inference: string | null;
  /** @minLength 1 */
  evidenceId: string;
  provenanceType: EvidenceItemProvenanceType;
  confirmed?: boolean;
  /** @minLength 1 */
  sourceLabel?: string;
}

export type EvidenceArray = EvidenceItem[];

export interface EvidenceCategories {
  scientificNeeds: EvidenceArray;
  workflows: EvidenceArray;
  samples: EvidenceArray;
  technologies: EvidenceArray;
  translationalStage: EvidenceArray;
  negativeOrContradictoryEvidence: EvidenceArray;
  materialUnknowns: EvidenceArray;
  buyingReadinessSignals: EvidenceArray;
}

export interface InstrumentEvidence {
  cellScapeRelevant: EvidenceArray;
  cosMxRelevant: EvidenceArray;
  geoMxRelevant: EvidenceArray;
}

export interface QualificationEvidence {
  schemaVersion: '1.0';
  purpose: 'EVIDENCE_ONLY_NO_INSTRUMENT_SELECTION';
  generationStatus: QualificationEvidenceGenerationStatus;
  categories: EvidenceCategories;
  instrumentDiscriminatingEvidence: InstrumentEvidence;
}

export interface ResearchPacket {
  schemaVersion: 'bsb-company-research-v1';
  /** @minLength 1 */
  brief: string;
  qualificationEvidence: QualificationEvidence;
}

export interface ResearchPacketRequest {
  researchPacket: ResearchPacket;
}

export interface PacketSummary {
  id: string;
  stage: string;
  brief: string;
  createdAt: string;
}

export type PacketRecordStage = typeof PacketRecordStage[keyof typeof PacketRecordStage];


export const PacketRecordStage = {
  VALIDATED: 'VALIDATED',
  NEEDS_REVIEW: 'NEEDS_REVIEW',
  ASSESSED: 'ASSESSED',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;

export type NormalizedEvidenceSupportStatus = typeof NormalizedEvidenceSupportStatus[keyof typeof NormalizedEvidenceSupportStatus];


export const NormalizedEvidenceSupportStatus = {
  SUPPORTED: 'SUPPORTED',
  UNSUPPORTED: 'UNSUPPORTED',
  NOT_APPLICABLE: 'NOT_APPLICABLE',
} as const;

export type NormalizedEvidence = EvidenceItem & {
  locations: string[];
  supportStatus: NormalizedEvidenceSupportStatus;
  supportIssues?: string[];
};

export interface FieldIssue {
  path: string;
  message: string;
}

export interface ValidationResult {
  structurallyValid: boolean;
  supportValid: boolean;
  errors: FieldIssue[];
  warnings: FieldIssue[];
}

export type InstrumentAssessmentInstrument = typeof InstrumentAssessmentInstrument[keyof typeof InstrumentAssessmentInstrument];


export const InstrumentAssessmentInstrument = {
  CellScape: 'CellScape',
  CosMx: 'CosMx',
  GeoMx: 'GeoMx',
} as const;

export type InstrumentAssessmentFit = typeof InstrumentAssessmentFit[keyof typeof InstrumentAssessmentFit];


export const InstrumentAssessmentFit = {
  STRONG_FIT: 'STRONG_FIT',
  POTENTIAL_FIT: 'POTENTIAL_FIT',
  NOT_QUALIFIED: 'NOT_QUALIFIED',
  INSUFFICIENT_EVIDENCE: 'INSUFFICIENT_EVIDENCE',
} as const;

export interface InstrumentAssessment {
  instrument: InstrumentAssessmentInstrument;
  fit: InstrumentAssessmentFit;
  recommendation: string;
  evidenceIds: string[];
  ruleIds: string[];
  alternatives: string[];
  currentUse: string;
  accountStatus: string;
  readiness: string;
}

export interface Assessment {
  id: string;
  provider: 'DETERMINISTIC_FAKE';
  mock: true;
  evidenceVersion: string;
  /** @maxItems 2 */
  instruments: InstrumentAssessment[];
  limitations: string[];
}

export type ReviewDecision = typeof ReviewDecision[keyof typeof ReviewDecision];


export const ReviewDecision = {
  APPROVE: 'APPROVE',
  REJECT: 'REJECT',
} as const;

export interface Review {
  id: string;
  decision: ReviewDecision;
  /** @maxItems 2 */
  approvedInstruments: string[];
  evidenceVersion: string;
  createdAt: string;
}

export interface PacketRecord {
  id: string;
  stage: PacketRecordStage;
  inputHash: string;
  researchPacket: ResearchPacket;
  normalizedEvidence: NormalizedEvidence[];
  validation: ValidationResult;
  assessment?: Assessment;
  review?: Review;
  createdAt: string;
}

export interface ValidationFailure {
  error: string;
  issues: FieldIssue[];
}

export type ReviewInputDecision = typeof ReviewInputDecision[keyof typeof ReviewInputDecision];


export const ReviewInputDecision = {
  APPROVE: 'APPROVE',
  REJECT: 'REJECT',
} as const;

export type ReviewInputApprovedInstrumentsItem = typeof ReviewInputApprovedInstrumentsItem[keyof typeof ReviewInputApprovedInstrumentsItem];


export const ReviewInputApprovedInstrumentsItem = {
  CellScape: 'CellScape',
  CosMx: 'CosMx',
  GeoMx: 'GeoMx',
} as const;

export interface ReviewInput {
  assessmentId: string;
  evidenceVersion: string;
  decision: ReviewInputDecision;
  /** @maxItems 2 */
  approvedInstruments: ReviewInputApprovedInstrumentsItem[];
  note: string;
}