# lib/api-zod/src/generated/types/evidenceCategories.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceArray } from './evidenceArray';

export interface EvidenceCategories {
  scientificNeeds: EvidenceArray;
  workflows: EvidenceArray;
  samples: EvidenceArray;
  technologies: EvidenceArray;
  translationalStage: EvidenceArray;
  negativeOrContradictoryEvidence: EvidenceArray;
  materialUnknowns: EvidenceArray;
  buyingReadinessSignals: EvidenceArray;
}