# lib/api-zod/src/generated/types/evidenceItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceItemAssessmentType } from './evidenceItemAssessmentType';
import type { EvidenceItemEvidenceState } from './evidenceItemEvidenceState';
import type { EvidenceItemProvenanceType } from './evidenceItemProvenanceType';

export interface EvidenceItem {
  evidenceState: EvidenceItemEvidenceState;
  assessmentType: EvidenceItemAssessmentType;
  /** @minLength 1 */
  claim: string;
  /**
     * @nullable
     * @pattern ^https?://
     */
  sourceUrl: string | null;
  basisFacts: string[];
  /** @items.pattern ^https?:// */
  basisSourceUrls: string[];
  /** @nullable */
  inference: string | null;
  /** @minLength 1 */
  evidenceId: string;
  provenanceType: EvidenceItemProvenanceType;
  confirmed?: boolean;
  /** @minLength 1 */
  sourceLabel?: string;
}