# lib/api-zod/src/generated/types/instrumentAssessment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { InstrumentAssessmentFit } from './instrumentAssessmentFit';
import type { InstrumentAssessmentInstrument } from './instrumentAssessmentInstrument';

export interface InstrumentAssessment {
  instrument: InstrumentAssessmentInstrument;
  fit: InstrumentAssessmentFit;
  recommendation: string;
  evidenceIds: string[];
  ruleIds: string[];
  alternatives: string[];
  currentUse: string;
  accountStatus: string;
  readiness: string;
}