# lib/api-zod/src/generated/types/packetRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { Assessment } from './assessment';
import type { NormalizedEvidence } from './normalizedEvidence';
import type { PacketRecordStage } from './packetRecordStage';
import type { ResearchPacket } from './researchPacket';
import type { Review } from './review';
import type { ValidationResult } from './validationResult';

export interface PacketRecord {
  id: string;
  stage: PacketRecordStage;
  inputHash: string;
  researchPacket: ResearchPacket;
  normalizedEvidence: NormalizedEvidence[];
  validation: ValidationResult;
  assessment?: Assessment;
  review?: Review;
  createdAt: Date;
}