# lib/api-zod/src/generated/types/packetSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */

export interface PacketSummary {
  id: string;
  stage: string;
  brief: string;
  createdAt: Date;
}