# lib/api-zod/src/generated/types/researchPacket.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { QualificationEvidence } from './qualificationEvidence';

export interface ResearchPacket {
  schemaVersion: 'bsb-company-research-v1';
  /** @minLength 1 */
  brief: string;
  qualificationEvidence: QualificationEvidence;
}