# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewDecision } from './reviewDecision';

export interface Review {
  id: string;
  decision: ReviewDecision;
  /** @maxItems 2 */
  approvedInstruments: string[];
  evidenceVersion: string;
  createdAt: Date;
}