# lib/api-zod/src/generated/types/validationFailure.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BSB Sequence Generator V2 Phase 1 API
 * OpenAPI spec version: 0.1.0
 */
import type { FieldIssue } from './fieldIssue';

export interface ValidationFailure {
  error: string;
  issues: FieldIssue[];
}